move some concepts to a schemas module
